skills/sales-pipeline/requirements.txt
# Core dependencies
requests>=2.28.0

# For ICP Learning Analyzer (PostgreSQL connection)
# Only needed if using icp_learning_analyzer.py
psycopg2-binary>=2.9.0

# Optional: python-dotenv for loading .env files
python-dotenv>=1.0.0
